requirements.txt
django==1.10
django-ckeditor
pillow
django-cleanup
django-crispy-forms
django-recaptcha
    